move table to own component
